+++ docs/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0 (2024-03-17)
+
+**Incompatible changes**
+
+- changed `rpc.call` and `rpc.call_async` to return bytes instead of python
+  dicts/lists
+
+
 ## 0.1.0
 
 Initial version.